+++ src/components/calcite-dropdown/calcite-dropdown.tsx
@@ -8,13 +8,12 @@ import {
   Listen,
   Prop,
 } from "@stencil/core";
-import { focusElement } from "../../utils/dom";
 import {
   GroupRegistration,
   ItemKeyboardEvent,
 } from "../../interfaces/Dropdown";
 import { getKey } from "../../utils/key";
-import { getElementDir } from "../../utils/dom";
+import { focusElement, getElementDir } from "../../utils/dom";
 
 @Component({
   tag: "calcite-dropdown",
@@ -65,6 +64,13 @@ export class CalciteDropdown {
 
   /** specify whether the dropdown is opened by hover or click of the trigger element */
   @Prop({ mutable: true, reflect: true }) type: "hover" | "click" = "click";
+
+  /**
+  allow the dropdown to remain open after a selection is made
+  if the selection-mode of the selected item's containing group is "none", the dropdown will alwqys close
+  */
+
+  @Prop({ mutable: true, reflect: true }) disableCloseOnSelect: boolean = false;
 
   //--------------------------------------------------------------------------
   //
@@ -140,6 +146,9 @@ export class CalciteDropdown {
   /** fires when a dropdown item has been selected or deselected **/
   @Event() calciteDropdownSelect: EventEmitter<void>;
 
+  /** fires when a dropdown has been opened **/
+  @Event() calciteDropdownOpen: EventEmitter<void>;
+
   @Listen("click") openDropdown(e) {
     if (e.target === this.trigger || this.trigger.contains(e.target)) {
       e.preventDefault();
@@ -149,13 +158,22 @@ export class CalciteDropdown {
   }
 
   @Listen("click", { target: "window" }) closeCalciteDropdownOnClick(e) {
-    if (this.active && e.target.offsetParent !== this.el) {
+    if (
+      this.active &&
+      e.target.nodeName !== "CALCITE-DROPDOWN-ITEM" &&
+      e.target.nodeName !== "CALCITE-DROPDOWN-GROUP"
+    ) {
       this.closeCalciteDropdown();
     }
   }
 
   @Listen("calciteDropdownClose") closeCalciteDropdownOnEvent() {
     this.closeCalciteDropdown();
+  }
+
+  @Listen("calciteDropdownOpen", { target: "window" })
+  closeCalciteDropdownOnOpenEvent(e) {
+    if (e.target !== this.el) this.active = false;
   }
 
   @Listen("keydown") keyDownHandler(e) {
@@ -232,6 +250,11 @@ export class CalciteDropdown {
     this.updateSelectedItems();
     event.stopPropagation();
     this.calciteDropdownSelect.emit();
+    if (
+      !this.disableCloseOnSelect ||
+      event.detail.requestedDropdownGroupMode === "none"
+    )
+      this.closeCalciteDropdown();
   }
 
   @Listen("calciteDropdownGroupRegister") registerCalciteDropdownGroup(
@@ -282,7 +305,6 @@ export class CalciteDropdown {
         "calcite-dropdown-item"
       )
     );
-
     this.selectedItems = items.filter((item) => item.active);
   }
 
@@ -309,7 +331,7 @@ export class CalciteDropdown {
 
   private closeCalciteDropdown() {
     this.active = false;
-    this.trigger.focus();
+    focusElement(this.trigger);
   }
 
   private focusOnFirstActiveOrFirstItem(): void {
@@ -366,5 +388,6 @@ export class CalciteDropdown {
         animationDelayInMs
       );
     }
+    this.calciteDropdownOpen.emit();
   }
 }